+++ flutter_google_places_sdk_platform_interface/lib/flutter_google_places_sdk_platform.dart
@@ -9,18 +9,19 @@ abstract class FlutterGooglePlacesSdkPlatform extends PlatformInterface {
 
   /// "Powered by google" image that should be used when background is white
   static const AssetImage ASSET_POWERED_BY_GOOGLE_ON_WHITE = AssetImage(
-      'assets/google/powered_by_google_on_white.png',
-      package: 'flutter_google_places_sdk_platform_interface');
+    'assets/google/powered_by_google_on_white.png',
+    package: 'flutter_google_places_sdk_platform_interface',
+  );
 
   /// "Powered by google" image that should be used when background is not white
   static const AssetImage ASSET_POWERED_BY_GOOGLE_ON_NON_WHITE = AssetImage(
-      'assets/google/powered_by_google_on_non_white.png',
-      package: 'flutter_google_places_sdk_platform_interface');
+    'assets/google/powered_by_google_on_non_white.png',
+    package: 'flutter_google_places_sdk_platform_interface',
+  );
 
   static final Object _token = Object();
 
-  static FlutterGooglePlacesSdkPlatform _instance =
-      FlutterGooglePlacesSdkMethodChannel();
+  static FlutterGooglePlacesSdkPlatform _instance = FlutterGooglePlacesSdkMethodChannel();
 
   /// Singleton instance to the platform
   static FlutterGooglePlacesSdkPlatform get instance => _instance;
@@ -40,13 +41,13 @@ abstract class FlutterGooglePlacesSdkPlatform extends PlatformInterface {
     throw UnimplementedError('deinitialize() has not been implemented.');
   }
 
-  /// Initializes Places for the given application context with the given API key.
+  /// Initializes Places for the given application context with the given API key. Use [useNewApi] to enable the Places API (New) in the SDK
   ///
   /// All Places API responses are localized using the device's locale.
   /// This method should only be called once prior to using the Places API.
   /// You may call this method again to update the API key used;
   /// if so, all widgets and instances of PlacesClient will now use this new key.
-  Future<void> initialize(String apiKey, {Locale? locale}) {
+  Future<void> initialize(String apiKey, {Locale? locale, bool? useNewApi}) {
     throw UnimplementedError('initialize() has not been implemented.');
   }
 
@@ -56,7 +57,7 @@ abstract class FlutterGooglePlacesSdkPlatform extends PlatformInterface {
   }
 
   /// Updates the settings of the places client with the given API key and locale.
-  Future<void> updateSettings(String apiKey, {Locale? locale}) {
+  Future<void> updateSettings(String apiKey, {Locale? locale, bool? useNewApi}) {
     throw UnimplementedError('initialize() has not been implemented.');
   }
 
@@ -85,8 +86,7 @@ abstract class FlutterGooglePlacesSdkPlatform extends PlatformInterface {
     LatLngBounds? locationBias,
     LatLngBounds? locationRestriction,
   }) {
-    throw UnimplementedError(
-        'findAutocompletePredictions() has not been implemented.');
+    throw UnimplementedError('findAutocompletePredictions() has not been implemented.');
   }
 
   /// Fetches the details of a place.
@@ -116,5 +116,54 @@ abstract class FlutterGooglePlacesSdkPlatform extends PlatformInterface {
     int? maxHeight,
   }) {
     throw UnimplementedError('fetchPlacePhoto() has not been implemented.');
+  }
+
+  /// Fetches places based on an ambiguous text query.
+  ///
+  /// Only the requested [fields] will be returned. If none specified,
+  /// all fields will be returned.
+  ///
+  /// Note that different fields can incur different billing.
+  ///
+  /// For more info about billing: https://developers.google.com/maps/documentation/places/android-sdk/usage-and-billing#pricing-new
+  ///
+  /// For more info on text search: https://developers.google.com/maps/documentation/places/android-sdk/text-search
+  Future<SearchByTextResponse> searchByText(
+    String textQuery, {
+    required List<PlaceField> fields,
+    String? includedType,
+    int? maxResultCount,
+    LatLngBounds? locationBias,
+    LatLngBounds? locationRestriction,
+    double? minRating,
+    bool? openNow,
+    List<int>? priceLevels,
+    TextSearchRankPreference? rankPreference,
+    String? regionCode,
+    bool? strictTypeFiltering,
+  }) {
+    throw UnimplementedError('fetchPlacePhoto() has not been implemented.');
+  }
+
+  /// Search for place(s) of interest using a location.
+  ///
+  /// Only the requested [fields] will be returned. If none specified,
+  /// all fields will be returned.
+  ///
+  /// Note that different fields can incur different billing.
+  ///
+  /// For more info on nearby search: https://developers.google.com/maps/documentation/places/android-sdk/nearby-search
+  Future<SearchNearbyResponse> searchNearby({
+    required List<PlaceField> fields,
+    required CircularBounds locationRestriction,
+    List<String>? includedTypes,
+    List<String>? includedPrimaryTypes,
+    List<String>? excludedTypes,
+    List<String>? excludedPrimaryTypes,
+    NearbySearchRankPreference? rankPreference,
+    String? regionCode,
+    int? maxResultCount,
+  }) {
+    throw UnimplementedError('searchNearby() has not been implemented.');
   }
 }